Move argo template to be under src/
