+++ frontend/src/lib/CompareUtils.ts
@@ -16,7 +16,7 @@
 
 import { ApiRunDetail, ApiRun } from '../apis/run';
 import { CompareTableProps } from '../components/CompareTable';
-import { Workflow } from 'third_party/argo-ui/argo_template';
+import { Workflow } from 'src/third_party/mlmd/argo_template';
 import { chain, flatten } from 'lodash';
 import WorkflowParser from './WorkflowParser';
 import { logger } from './Utils';
