+++ frontend/src/lib/ParserUtils.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { Metadata, Workflow } from 'third_party/argo-ui/argo_template';
+import { Metadata, Workflow } from 'src/third_party/mlmd/argo_template';
 
 export function parseTaskDisplayName(metadata?: Metadata): string | undefined {
   if (!metadata?.annotations) {
