+++ frontend/src/lib/RunUtils.ts
@@ -15,7 +15,7 @@
  */
 
 import { ApiParameter, ApiPipelineVersion } from '../apis/pipeline';
-import { Workflow } from '../../third_party/argo-ui/argo_template';
+import { Workflow } from '../third_party/mlmd/argo_template';
 import { ApiJob } from '../apis/job';
 import {
   ApiPipelineRuntime,
