+++ frontend/src/pages/PipelineDetails.tsx
@@ -31,7 +31,7 @@ import {
 import * as WorkflowUtils from 'src/lib/v2/WorkflowUtils';
 import { convertJsonToV2PipelineSpec } from 'src/lib/v2/WorkflowUtils';
 import { classes } from 'typestyle';
-import { Workflow } from '../../third_party/argo-ui/argo_template';
+import { Workflow } from '../third_party/mlmd/argo_template';
 import { ApiExperiment } from '../apis/experiment';
 import { ApiGetTemplateResponse, ApiPipeline, ApiPipelineVersion } from '../apis/pipeline';
 import { QUERY_PARAMS, RoutePage, RouteParams } from '../components/Router';
